Updated serialization header generation
Due to the Vulkan XML being parsed newest to oldest, enum values that
were reserved with that keyword in the name would be parsed after the
replacement value and name, and thus be generated after in the enum set
lists.

This meant that when serializing, the 'RESERVED' version would almost
always be used instead of the replacement name. This was because the
serialization function would go in reverse order in attempts to use
larger combination values typically later in the enum list.

To try to use the 'proper' names instead when they are available, when
generating the header, if will now iterate through the available
versions and write out only the enums that appeared in that version.
This would lead to items added to the spec after to appear after, and
thus be used for serialization instead.