Changed order of comparison for VkResult strings
As extensions mature over time, the same value can be associated with
multiple names, perhaps originally as an 'EXT' then a 'KHR' extension.

To provide back the most accurate text for corresponding VkResult error
code values, the checking of codes introduced in later Vulkan versions
has been shifted to be first, and comparisons to be in descending order
that they appeared.