Added new 'nullAllowed' field for serialization
As has happened with the VkPipelineDepthStencilStateCreateFlagBits type,
enum sets that were once empty may change in the future, and at that
point not have a zero-value.

This breaks backward compatability, and so a new field has been added
that allows for empty fields/values if the type was once empty.