Value serialization now accounts for aliases
When serializing, the order that values were parsed could have an affect
on what is output as part of serializing values. The last value that
matched the value would always be used, even if that value was an
aliased value to a possibly promoted newer value.

To deal with this, the sets of enum values now have a new 'alias'
boolean which describes whether the value is aliased, and if so, will be
skipped for serialization. Per value, there is only ever one non-aliased
value, from which all the 'aliased' versions copy that value.

The reason these aliased values still exist in the sets is still for
parsing from data and programs using older versions of the Vulkan
standard.