Fix issues with type names with numbers in them
Some newer types, such as VkPipelineStageFlags2 has numeric values at
the end, leading to issues where the assumption of 'Flags' or 'FlagBits'
always being at the end no longer held true. This would lead to where
generated prefixes were incorrect, where the flags/bits were being added
to the prefix, and where the number values were being improperly added.

Instead of the desired prefix of:
VK_PIPELINE_STAGE_2_
it would generate this:
VK_PIPELINE_STAGE_FLAG_BITS2_

Thus, a new method of detecting the Flag/Bits using strstr, and instead
of manipulating the length of the string, it would advance the character
pointer when it reaches it.

As well, adding the case of inserting an underscore when a numeric value
is detected.