Fix issue when dealing with removed struct members
Through the life of the Vulkan spec, sometimes structures have their
sType members removed, leading to errors when doing sType comparisons
for the dynamic struct cleanup calls.

This changes generation to take into account the first and last time the
sType was seen on a member, and guard based on that for the dynamic
struct cleanup.