Rework parsing/generation for new APIs
The Vulkan spec XML file now holds two simultaneous APIs, the regular
Vulkan API and the new VulkanSC API.

As such, changes have been made to support this.

As well, better support for determining when types are available for
when a type is available only when multiple platforms/definitions are
available.