Fix struct alias issues with cleanup/compare calls
When a structure is an alias, the current implementation leaves an
otherwise empty function, when it really should be calling the new alias
version of the call.

Comparison functions will now redirect to the aliast version of the
call, with appropriate casts.

Cleanup functions no longer have inline declared functions. Instead, all
functions have definitions, even if the definition is otherwise empty.